refactor(ui): extract cosmic chrome to @amiclaw/ui workspace package
Lift 8 platform-shared chrome components from packages/game/src/components/
to a new packages/ui/ workspace package, plus AMIO platform tokens and
atlas animation keyframes. Both game (BombSquad) and game-yijing now import
from @amiclaw/ui, removing the duplicate-or-rebuild path before yijing's
5 hi-fi screens land.

Lifted to packages/ui/:
  - Scenery, Button, GlassCard, EyebrowTag, ConicAvatar, Chip,
    SectionHeader (preserves default + named accentClass), StatPill
  - styles/tokens.css (AMIO atlas palette + cosmic-shell aliases)
  - styles/animations.css (twinkle / spin / pulse / float keyframes
    plus prefers-reduced-motion neutralization)

Left in packages/game/:
  - PlatformLayout (composes claw-platform-specific nav chrome --
    TopNav, BottomNav, PlatformFooter -- that yijing does not consume;
    its Scenery import now resolves through @amiclaw/ui)
  - BombSquad-only styles/bombsquad-tokens.css (trimmed to game radii)
    and styles/animations.css (success-flash / error-flash / wire-cut-*)
  - All components/home/, components/bombsquad/, and game-engine code

Both consumer apps gained "@amiclaw/ui": "workspace:*" deps; main.tsx in
each loads ui/tokens.css + ui/animations.css before any consumer-local
CSS so the cascade resolves correctly.

Verified: pnpm typecheck clean for all 5 packages (api, ui, game,
game-yijing, manual); pnpm test:run 318/318 green (api 29 + manual 28 +
game 261 + game-yijing 0 + ui 0 -- BombSquad regression net intact). Zero
JS animation library introductions per amiclaw CSS-only invariant.

Source task: pages/projects/amiclaw/tasks/build-yijing-oracle-screens
(Round 1 of the sibling-2 plan; subsequent rounds implement the 5 yijing
hi-fi screens consuming this shared chrome.)
