+++ packages/game-yijing/src/main.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
+import '@amiclaw/ui/styles/tokens.css'
+import '@amiclaw/ui/styles/animations.css'
 import './styles/tokens.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
