+++ packages/game/package.json
@@ -10,6 +10,7 @@
     "test:run": "vitest run"
   },
   "dependencies": {
+    "@amiclaw/ui": "workspace:*",
     "js-yaml": "^4.1.0",
     "react": "^19.2.6",
     "react-dom": "^19.2.6",
