+++ packages/game/src/main.tsx
@@ -3,11 +3,11 @@ import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 import App from './App'
 import './styles/fonts.css'
-import './styles/atlas-tokens.css'
+import '@amiclaw/ui/styles/tokens.css'
+import '@amiclaw/ui/styles/animations.css'
 import './styles/bombsquad-tokens.css'
 import './styles/global.css'
 import './styles/animations.css'
-import './styles/atlas-animations.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
