+++ packages/game/src/pages/ConnectPage.tsx
@@ -1,9 +1,8 @@
 import { useEffect, useState } from 'react'
 import { Link, useNavigate, useSearchParams } from 'react-router-dom'
-import Scenery from '@/components/platform/Scenery'
+import { Scenery, ConicAvatar } from '@amiclaw/ui'
 import Eyebrow from '@/components/bombsquad/Eyebrow'
 import Button from '@/components/bombsquad/Button'
-import ConicAvatar from '@/components/ui/ConicAvatar'
 import { useDailyChallenge } from '@/hooks/useDailyChallenge'
 import { copyToClipboard } from '@/utils/clipboard'
 import styles from './ConnectPage.module.css'
