+++ packages/game/src/pages/GamePage.tsx
@@ -23,7 +23,7 @@ import SceneInfoBar from '@/components/SceneInfoBar'
 import MuteButton from '@/components/MuteButton'
 import StrikeIndicator from '@/components/StrikeIndicator'
 import ExplosionOverlay from '@/components/ExplosionOverlay'
-import Scenery from '@/components/platform/Scenery'
+import { Scenery } from '@amiclaw/ui'
 import Eyebrow from '@/components/bombsquad/Eyebrow'
 import WireModule from '@/modules/wire/WireModule'
 import DialModule from '@/modules/dial/DialModule'
