+++ packages/ui/package.json
@@ -1,28 +1,31 @@
 {
-  "name": "@amiclaw/game-yijing",
+  "name": "@amiclaw/ui",
   "private": true,
   "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./styles/tokens.css": "./src/styles/tokens.css",
+    "./styles/animations.css": "./src/styles/animations.css"
+  },
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
     "typecheck": "tsc --noEmit",
     "test": "vitest --passWithNoTests",
     "test:run": "vitest run --passWithNoTests"
   },
-  "dependencies": {
-    "@amiclaw/ui": "workspace:*",
+  "peerDependencies": {
     "react": "^19.2.6",
     "react-dom": "^19.2.6",
     "react-router-dom": "^7.15.1"
   },
   "devDependencies": {
-    "@types/node": "^25.9.1",
     "@types/react": "^19.2.15",
     "@types/react-dom": "^19.2.3",
-    "@vitejs/plugin-react": "^6.0.2",
-    "jsdom": "^29.1.1",
+    "react": "^19.2.6",
+    "react-dom": "^19.2.6",
+    "react-router-dom": "^7.15.1",
     "typescript": "^6.0.3",
-    "vite": "^8.0.14",
     "vitest": "^4.1.7"
   }
 }
