+++ packages/ui/src/GlassCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GlassCard'
