+++ packages/ui/src/Scenery/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Scenery'
