+++ packages/ui/src/SectionHeader/index.ts
@@ -1,0 +1,1 @@
+export { default, accentClass } from './SectionHeader'
