+++ packages/ui/src/StatPill/index.ts
@@ -1,0 +1,1 @@
+export { default } from './StatPill'
