+++ packages/ui/src/index.ts
@@ -1,0 +1,13 @@
+/* @amiclaw/ui — barrel export for cross-game cosmic shell chrome.
+   Each component preserves its original default-export shape; the
+   barrel re-exports defaults under their named identifiers so
+   consumers can `import { Button, ConicAvatar, ... } from '@amiclaw/ui'`. */
+
+export { default as Scenery } from './Scenery'
+export { default as Button } from './Button'
+export { default as GlassCard } from './GlassCard'
+export { default as EyebrowTag } from './EyebrowTag'
+export { default as ConicAvatar } from './ConicAvatar'
+export { default as Chip } from './Chip'
+export { default as SectionHeader, accentClass } from './SectionHeader'
+export { default as StatPill } from './StatPill'
