+++ packages/ui/src/styles/animations.css
